pms-be/requirements.txt
Django>=5.2.5, <5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt>=5.5.1, <5.6.0
django-cors-headers>=4.7.0, <4.8.0
psycopg2-binary>=2.9.10, <2.10.0
python-decouple>=3.8, <3.9.0
gunicorn>=23.0.0, <23.1.0